pyproject.toml
[build-system]
requires = ['setuptools >= 61.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.6.0",
    "certifi>=2024.8.30",
    "click>=8.1.7",
    "distro>=1.9.0",
    "h11>=0.14.0",
    "httpcore>=1.0.5",
    "httpx>=0.27.2",
    "idna>=3.10",
    "jiter>=0.5.0",
    "markdown-it-py>=3.0.0",
    "mdurl>=0.1.2",
    "openai>=1.47.0",
    "pydantic>=2.9.2",
    "pydantic_core>=2.23.4",
    "Pygments>=2.18.0",
    "python-dotenv>=1.0.1",
    "pyperclip>=1.9.0",
    "rich>=13.8.1",
    "sniffio>=1.3.1",
    "tqdm>=4.66.5",
    "typing_extensions>=4.12.2",
    "wrapworks>=1.0.6",
]
name = "pythion"
version = "1.0.22"
description = "AI-powered CLI tool for generating Python documentation strings effortlessly"
readme = "README.md"
requires-python = ">= 3.10"
authors = [{ name = "Asaniczka", email = "asaniczka@gmail.com" }]
maintainers = [{ name = "Asaniczka", email = "asaniczka@gmail.com" }]
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.10",
]

[project.urls]
Homepage = "https://github.com/asaniczka/pythion"


[project.scripts]
pythion = "pythion.__init__:pythion"
